# Speedup cpu memory write read