Optimizes `Cpu.read_bytes` and `Cpu.write_bytes`.

Tbh I am not sure if it does not breaks anything but from some of my reading I think not. Let's see during CI tests.

Benchmark below:
```python
def test_cpu_write_read():
    import time
    from manticore.core.smtlib import ConstraintSet
    from manticore.native.memory import SMemory64
    from manticore.native.cpu.x86 import AMD64Cpu

    cs = ConstraintSet()
    mem = SMemory64(cs)
    cpu = AMD64Cpu(mem)

    buf = mem.mmap(None, 0x1000, "rw")

    string = b'a' * 0x1000

    t1 = time.time()
    for i in range(1000):
        cpu.write_bytes(buf, string, 0x1000)
    t2 = time.time()

    t3 = time.time()
    for i in range(1000):
        cpu.read_bytes(buf, 0x1000)
    t4 = time.time()

    print(f"test cpu write_bytes: {t2-t1} seconds")
    print(f"test cpu read_bytes : {t4-t3} seconds")

    assert cpu.read_bytes(buf, 0x1000) == [b'a'] * 0x1000

test_cpu_write_read()
```

On master (01284c01c920bbfcb7c51991a447c7b850f6523f):
```
$ pip freeze | grep manticore
-e git+git@github.com:trailofbits/manticore.git@01284c01c920bbfcb7c51991a447c7b850f6523f#egg=manticore

(manticore) d~@lb:~/perfwrite$ python testperf.py
test cpu write_bytes: 74.96629548072815 seconds
test cpu read_bytes : 114.84957957267761 seconds
```

On this PR (c1aaa4a9a81f2a48c12cc1fb6d07389a5ec2ae96):
```
(manticore-perf) d~@lb:~/perfwrite$ pip freeze | grep manticore
-e git+git@github.com:trailofbits/manticore.git@c1aaa4a9a81f2a48c12cc1fb6d07389a5ec2ae96#egg=manticore

(manticore-perf) d~@lb:~/perfwrite$ python testperf.py
test cpu write_bytes: 46.61375951766968 seconds
test cpu read_bytes : 6.370912790298462 seconds
```

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/trailofbits/manticore/1510)
<!-- Reviewable:end -->