@disconnect3d 2019-08-27
Btw I am not sure about this PR yet: it fails some tests. I looked at them a bit and it seems it might be a result of some types mismatch (e.g. we used ORD/CHR to push values into the memory and now -- in this PR -- we don't).

I don't have time to dive into it now, so we can either close it for now and revisit later or keep it open.